+++ server/clearbit/utils.js
@@ -15,8 +15,12 @@ export function isInSegments(userSegments = [], segmentsListIds = []) {
   ).length > 0;
 }
 
-export function getDomain(user) {
-  return user["traits_clearbit/employment_domain"] || user["traits_clearbit_company/domain"] || user.domain;
+export function getDomain(user, account = {}) {
+  return user["traits_clearbit/employment_domain"]
+      || user["traits_clearbit_company/domain"]
+      || account.domain
+      || account["clearbit_company/domain"]
+      || user.domain;
 }
 
 export function now() {